all id fields are changed to be int64, and amount to be int32 and balance to be int64

+++ responses/account_responses.go
@@ -6,19 +6,19 @@ import (
 )
 
 type CreateAccountResponse struct {
-	AccountID uint64    `json:"account_id"`
+	AccountID int64     `json:"account_id"`
 	CreatedAt time.Time `json:"created_at"`
 	Owner     string    `json:"owner"`
-	Balance   uint64    `json:"balance"`
+	Balance   int64     `json:"balance"`
 }
 
 type GetAccountResponse struct {
-	AccountID uint64         `json:"account_id"`
+	AccountID int64          `json:"account_id"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"deleted_at"`
 	Owner     string         `json:"owner"`
-	Balance   uint64         `json:"balance"`
+	Balance   int64          `json:"balance"`
 }
 
 type ListAccountsResponse struct {
@@ -26,8 +26,8 @@ type ListAccountsResponse struct {
 }
 
 type EntryResponse struct {
-	EntryID   uint64    `json:"entry_id"`
-	AccountID uint64    `json:"account_id"`
+	EntryID   int64     `json:"entry_id"`
+	AccountID int64     `json:"account_id"`
 	CreatedAt time.Time `json:"created_at"`
-	Amount    int64     `json:"amount"`
+	Amount    int32     `json:"amount"`
 }
